feat(nub-arch-x86): remove the eager mem_buf — true zero-setup demand paging
The catch-all UnpinnedCapCow MatRange now covers the entire data extent (sourced
from inst.mem's slabs / the shared zero page), so every in-data page is
materialized lazily via mat_pas. The eager flat RW buffer has no remaining role:

- Delete FrameRuntime.mem_buf + mem_buf_pa, the MAT_MEM_BUF_PA static, and its
  publish in enter_frame. build_frame_runtime allocates only CTX + STACK.
- try_materialize: the ephemeral arm (mem_buf_pa + offset) is unreachable (a
  page with no MatRange can't be in-data) → fault. Collapse the former
  EphemeralZero map-writable special case into cow_into_fresh: every writable
  page CoWs from src_pa, which for an ephemeral page is the shared zero page
  (CoW-from-zero). Gas-neutral — charge_for never distinguished EphemeralZero
  from UnpinnedCapCow.

Also refresh the now-stale rw_overlays/data_overlays/mem_buf doc comments
across jit_run, nub-arch-local, and the conformance harness.

Gates re-run green: differential interp==recomp (15), 3-way conformance,
sub_vm, workloads (10); clippy --workspace + fmt clean.

+++ rust/javm-cap/src/image.rs
@@ -210,7 +210,7 @@ impl Image {
             if content.is_empty() {
                 continue;
             }
-            let base_off = (mapping.start as u64).saturating_sub(data_base);
+            let base_off = mapping.start.saturating_sub(data_base);
             for (i, chunk) in content.chunks(PAGE_SIZE).enumerate() {
                 backing.put_page(base_off + (i * PAGE_SIZE) as u64, chunk);
             }
